requirements.txt
# Django
Django>=5.2.9,<6.0

# Configuration et variables d'environnement
python-decouple>=3.8

# Base de données PostgreSQL
psycopg2-binary>=2.9.9

# Fichiers statiques (WhiteNoise pour o2switch)
whitenoise>=6.6.0

# Qualité de code
flake8>=7.0.0

# Tests et couverture
coverage>=7.5.0

# Utilitaires
python-dotenv>=1.0.0
Pillow>=10.0.0


